Updated performance notes. See [6d818bf1f32ac653] for comparison.
|Hash|Bytes   |Micros     |MB/Sec|MiB/Sec|Micros/Byte|Gain  |R/Gain|
|--- |---:    |---:       |---:  |---:   |---:       |---:  |---:  |
|md4 |  524288|    686.469|763.75| 728.37|0.0013     |11.967| 1.564|
|md5 | 1048576|   2146.825|488.43| 465.80|0.0020     | 7.653| 2.565|
|sha1|67108864| 352426.623|190.42| 181.60|0.0053     | 2.984| 1.094|
|sha2|  262144|   1505.720|174.10| 166.03|0.0057     | 2.728| 2.728|
|sha3|67108864|1051598.331| 63.82|  60.86|0.0157     | 1.000| 1.000|

With the exception of SHA-1 all hashes got a performance boost. The
oldest of the hashes, MD-4, got the largest boost, jumping it to the
top spot, overtaking both MD-5 and SHA-1.The modern SHA-3 is still
trailing all the others, although not as much anymore as before the
switch to RHash.

Overall I call that a success.