Dropped the input files for the test vectors. These are now all in the test utilities. See commit [d02c310d65].
Added the variant sha2/3 hashes to the set of tested functions.
In the utility functions.
Added the necessary result files.
Reworked hash.test to use the data from the utilities.

Ditto for references.

Not for list, to prevent it from failing. Switching is not possible
until the hash list in the utilities covers all the supported functions.